Added text-number mapping for first 10 digits in get_article_list_count() method
The method get_article_list_count() is no longer limited to numbers in digit format. If the article title starts with any number one through ten in text format, the method get_article_list_count() will now assume the article contains a list with the quantity equal to that number.